requirements.txt
accelerate==0.33.0
bitsandbytes==0.43.3
fastapi==0.109.0
gradio==4.37.2
loguru
matplotlib
numpy==1.26.3
pandas==2.2.0
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
# sentence-transformers
# torch==2.3.1+cpu
# torchaudio==2.3.1+cpu
# torchvision==0.18.1+cpu
# transformers==4.42.3
